% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_BPD_Class.R
\docType{data}
\name{bpd_cla}
\alias{bpd_cla}
\title{An OTU table of class level taxa counts from a BPD1 study}
\format{
A 34x75 data.frame
\describe{
  \item{\code{OTU_Name}}{A character vector of class level OTU names}
  \item{\code{Lib names}}{The following columns are the sequencing counts for each library with library names}
}
}
\source{
\url{https://doi.org/10.1371/journal.pone.0170120}
}
\usage{
bpd_cla
}
\description{
Infants who required mechanical ventilation had tracheal aspirates samples collected at 7, 14, and 21 days of age (+/- 48 hours). Infants who were mechanically ventilated and had at least one tracheal aspirate collected were included in this study. Subjects were required to be enrolled within 7 days of age. Bacterial profiles were determined by broad-range amplification and sequence analysis of 16S rRNA genes.
}
\keyword{datasets}
