% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_BPD_Family.R
\docType{data}
\name{bpd_fam}
\alias{bpd_fam}
\title{An OTU table of family level taxa counts from BPD1 study}
\format{
A 116x75 data.frame
\describe{
\item{\code{OTU_Name}}{A character vector of family level OTU names}
\item{\code{Lib names}}{The following columns are the sequencing counts for each library with library names}
}
}
\source{
\url{https://doi.org/10.1371/journal.pone.0170120}
}
\usage{
bpd_fam
}
\description{
Infants who required mechanical ventilation had tracheal aspirates samples collected at 7, 14, and 21 days of age (+/- 48 hours). Infants who were mechanically ventilated and had at least one tracheal aspirate collected were included in this study. Subjects were required to be enrolled within 7 days of age. Bacterial profiles were determined by broad-range amplification and sequence analysis of 16S rRNA genes.
}
\keyword{datasets}
