% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_MRSA_Gen.R
\docType{data}
\name{mrsa_gen}
\alias{mrsa_gen}
\title{An OTU table of genus level taxa counts from a MRSA study}
\format{
A 34x75 data.frame
\describe{
  \item{\code{OTU_Name}}{A character vector of class level OTU names}
  \item{\code{Lib names}}{The following columns are the sequencing counts for each library with library names}
}
}
\source{
\url{http://dx.doi.org/10.1016/j.jinf.2015.08.008}
}
\usage{
mrsa_gen
}
\description{
Data from a study to define the nasal microbiome of hospital inpatients who are persistently colonized with methicillin-resistant Staphylococcus aureus (MRSA) compared with matched, non-colonized controls. See original paper for matching proceedure.
}
\keyword{datasets}
